Added data persistence
The score of the current game is persisted between scenes.

Every score at the end of the game is saved if the player enters a name. The highest of these scores is displayed to the player during the game.

All scores are being persisted so a more complex leaderboard can be added later.